+++ apps/cowswap-frontend/src/modules/tradeQuote/state/tradeQuoteAtom.ts
@@ -4,6 +4,7 @@ import { getCurrencyAddress } from '@cowprotocol/common-utils'
 import { getAddressKey, QuoteAndPost } from '@cowprotocol/cow-sdk'
 import { BridgeProviderQuoteError, BridgeQuoteResults } from '@cowprotocol/sdk-bridging'
 
+import { isProviderNetworkDeprecatedAtom } from 'entities/common/isProviderNetworkDeprecated.atom'
 import { isProviderNetworkUnsupportedAtom } from 'entities/common/isProviderNetworkUnsupported.atom'
 
 import { derivedTradeStateAtom } from 'modules/trade/state/derivedTradeStateAtom'
@@ -73,13 +74,14 @@ export const updateTradeQuoteAtom = atom(
 
 export const currentTradeQuoteAtom = atom<TradeQuoteState>((get) => {
   const isProviderNetworkUnsupported = get(isProviderNetworkUnsupportedAtom)
+  const isProviderNetworkDeprecated = get(isProviderNetworkDeprecatedAtom)
   const state = get(derivedTradeStateAtom)
   const tradeQuotes = get(tradeQuotesAtom)
 
   const inputCurrency = state?.inputCurrency
   const outputCurrency = state?.outputCurrency
 
-  if (!inputCurrency || !outputCurrency || isProviderNetworkUnsupported === true) {
+  if (!inputCurrency || !outputCurrency || isProviderNetworkUnsupported || isProviderNetworkDeprecated) {
     return DEFAULT_TRADE_QUOTE_STATE
   }